+++ tests/models/cassettes/test_openai_responses/test_openai_responses_model_thinking_part.yaml
